+++ lib/comparator.dart
@@ -2,16 +2,17 @@
 /// methods to modify and perform common operations on [Comparator] functions.
 library;
 
+export 'src/comparator/compare_extensions.dart';
 export 'src/comparator/compare_operators.dart';
-export 'src/comparator/constructors/delegate.dart';
 export 'src/comparator/constructors/explicit.dart';
+export 'src/comparator/constructors/key_of.dart';
 export 'src/comparator/constructors/natural.dart';
 export 'src/comparator/constructors/reverse.dart';
 export 'src/comparator/modifiers/compound.dart';
+export 'src/comparator/modifiers/key_of.dart';
 export 'src/comparator/modifiers/lexicographical.dart';
 export 'src/comparator/modifiers/nulls_first.dart';
 export 'src/comparator/modifiers/nulls_last.dart';
-export 'src/comparator/modifiers/result_of.dart';
 export 'src/comparator/modifiers/reversed.dart';
 export 'src/comparator/operations/binary_search.dart';
 export 'src/comparator/operations/largest.dart';