+++ lib/src/collection/string/chunked.dart
@@ -2,6 +2,8 @@ import 'dart:math';
 
 import 'package:characters/characters.dart';
 
+import '../../shared/exceptions.dart';
+
 extension ChunkedStringExtension on String {
   /// Divides this [String] into an iterable of strings each not exceeding the
   /// given [size]. The last string might have fewer characters.
@@ -13,6 +15,7 @@ extension ChunkedStringExtension on String {
   /// print(input.chunked(2));  // ('ab', 'cd', 'e')
   /// ```
   Iterable<String> chunked(int size) sync* {
+    checkNonZeroPositive(size);
     for (var i = 0; i < length; i += size) {
       yield substring(i, min(i + size, length));
     }
@@ -30,6 +33,7 @@ extension ChunkedCharactersExtension on Characters {
   /// print(input.chunked(2));  // ('ab', 'cd', 'e')
   /// ```
   Iterable<Characters> chunked(int size) sync* {
+    checkNonZeroPositive(size);
     final range = iterator;
     while (range.moveNext(size)) {
       yield range.currentCharacters;