+++ lib/src/diff/close_matches.dart
@@ -28,10 +28,9 @@ extension CloseMatchesOnIterable<T> on Iterable<Iterable<T>> {
     return candidates
         .largest(
           count,
-          comparator:
-              delegateComparator<({Iterable<T> candidate, double ratio}), num>(
-                (each) => each.ratio,
-              ),
+          comparator: keyOf<({Iterable<T> candidate, double ratio}), num>(
+            (each) => each.ratio,
+          ),
         )
         .map((each) => each.candidate);
   }